% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineFunctions.R
\name{additionalEggsProcessing}
\alias{additionalEggsProcessing}
\title{Processing for egg calculations}
\usage{
additionalEggsProcessing(fec)
}
\value{
List of number of eggs spawned in each production unit
by simulated females, or the carrying capacity: whichever
is smaller.
}
\description{
Internal function used to sum number of eggs in each 
production unit from each of the four migration
routes, and apply carrying capacity based on k_pus. In addition
to cohort processing routines (hence the name). Not
intended to be called directly. Visible for the sake of
model transparency.
}
