% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualUpstream.R
\name{annualUpstream}
\alias{annualUpstream}
\title{Annual upstream migration parameters}
\usage{
annualUpstream(river, damRkms)
}
\arguments{
\item{NULL}{}
}
\value{
A named list of objects used for the annual 
upstream migration model, including 1) upstream dam 
passage efficiencies for each of the four possible 
migration routes (list), 2) probability of using the 
mainstem Penobscot River around Marsh Island (vector),
3) the probability of using the mainstem Penobscot 
River during downstream migration (vector), and
4) probability of using the mainstem Penobscot River
at the confluence with the Piscataquis River (vector).
}
\description{
A helper function used to define upstream passage efficiencies 
and routes based on stochastic variables and user-defined
inputs to \code{\link{penobscotRiverModel}}, as well as
the probability of using the mainstem Penobscot River
during downstream migration. Not intended to
be called directly, but visible for model transparency.
}
