% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arr.B}
\alias{arr.B}
\title{Arrival regression parameters for bucks}
\format{
A list of 1000 matrices with named dimensions of:
\describe{
  \code{dim1 }{Intercept, slope}

  \code{dim2 }{Estimate, Std Error, t statistic, and p-value}
}
}
\source{
Connecticut Department of Energy and Environmental Conservation (2015)
}
\usage{
arr.B
}
\description{
Bootstrapped regression parameters for
a quasibinomial logistic regression used to predict
arrival timing of male American shad in the Connecticut
River, CT, USA by accumulated thermal units (ATU).
}
\keyword{datasets}
