% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignFishToPU.R
\name{assignFishToPU}
\alias{assignFishToPU}
\title{Assign fish to routes}
\usage{
assignFishToPU(puNum, PUS)
}
\description{
Generically generate individual x_PUS_n,
for fish gender and PU number. Not intended to be called
directly. Visible for the sake of model transparency.
}
