% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignFishToRoutes.R
\name{assignFishToRoutes}
\alias{assignFishToRoutes}
\title{Assign fish to routes}
\usage{
assignFishToRoutes(river, c_fishAges)
}
\description{
River-specific assignment to migration routes for
individual fish.

Internal function used inside inner-loop-sampling, but stored in
a separate file to make code easier to work with.
}
