% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b.parms_merrimack}
\alias{b.parms_merrimack}
\title{von Bertalanffy growth parameters for roes}
\format{
A list of 10000 matrices with named dimensions
\describe{
  \code{dim1 }{Sinf, K, t0}

  \code{dim2 }{Estimate, Std Error, t statistic, and p-value}
}
}
\source{
U.S. Fish and Wildlife Service (2018)
}
\usage{
b.parms_merrimack
}
\description{
Bootstrapped regression parameters for
von Bertalanffy growth parameters for models fit to
male American shad in the Merrimack River, USA. Uses
data from fish < 3 years of age from the Connecticut
River, USA to calibrate growth curves.
}
\keyword{datasets}
