% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineFunctions.R
\name{CI}
\alias{CI}
\title{Calculate 95 percent confidence intervals}
\usage{
CI(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Returns the 2.5th and 97.5th quantiles of
an object.
}
\description{
A helper function to calculate 95% confidence
intervals on an object.
}
\examples{

CI(rnorm(n=1e3, 0, 1))

}
