% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPUMatrix.R
\name{createPUMatrix}
\alias{createPUMatrix}
\title{Create production unit matrix}
\usage{
createPUMatrix(isFemale, pu, isEgg = FALSE)
}
\value{
Matrix of fish numbers by ages in each PU following
execution of the upstream migration model in source code.
}
\description{
Internal function to generalize creation of
the PU_n data structure for different PUs and fish sexes.
Not intended to be called directly, but visible
for model transparency.
}
