% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctr_proj85}
\alias{ctr_proj85}
\title{Projected daily river temperatures (RCP8.5) in CT River, USA}
\format{
A dataframe with 54,747 obs. of 5 variables:
\describe{
  \code{Date }{Date for projection}
  
  \code{Haddam }{Daily temperature (C) at Haddam}
  
  \code{Thompsonville }{Daily temperature (C) at Thompsonville}

  \code{Turners }{Daily temperature (C) at Turners Falls}
  
  \code{avg}{Average daily temperature across sites}
}
}
\source{
Integrated Systems Ecology Lab, Gulf Of Maine Research Institute
}
\usage{
ctr_proj85
}
\description{
Daily project temperatures in the Connecticut
River, USA for years 1950-2099 under RCP8.5. Three sites
are included based on availbility of time-series used to
construct the projections.
}
\keyword{datasets}
