% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineHabitat.R
\name{defineHabitat}
\alias{defineHabitat}
\title{Production potential definition}
\usage{
defineHabitat()
}
\value{
A list of habitat production potential
for production units in the Penobscot River, Maine, USA.
}
\description{
Internal function used to
assign initial values of production potential
for American shad in production units of the 
Penobscot River, Maine, USA. 

Not intended to be called directly, but visible 
for model transparency.
}
\references{
Maine Department of Marine Resources (MDMR). 2009. 
Operational plan for the restoration of diadromous fishes 
to the Penobscot River. Final Report, MDMR, 
Maine Department of Inland Fish and Wildlife, 
Augusta, ME. \url{http://maine.gov/dmr/searunfish/reports/Penobscot_Operational_Plan_final_2009.pdf}.
}
