% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineHydroSystem.R
\name{defineHydroSystem}
\alias{defineHydroSystem}
\title{Definition of hydro system}
\usage{
defineHydroSystem()
}
\value{
A list of values related to the hydro system, 
including 1) the number of migration routes, 2) number of 
dams, 3) number of production units delineated by dams,
4) the upstream extent to habitat in river kilometers (rkm), 
and 5) the locations of dams in rkm.
}
\description{
Internal function used to define the
configuration of the hydrosystem and catchment. Not
intended to be called directly, but visible for the
sake of model transparency.
}
