% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-defineOutputVectors-last.R
\name{defineOutputVectors_last}
\alias{defineOutputVectors_last}
\title{Define simulation output vectors and pre-allocate memory}
\usage{
defineOutputVectors_last()
}
\value{
A list of all stored simulation parameters (90
total, some unused but retained for transparency
in previous management strategy evaluations and
hydropower licensing negotiations).
}
\description{
Internal function to pre-allocate output
vectors to hold the results of simulation runs before
they are output to the global workspace.

Not intended to be called directly, but included
for model transparency.
}
