% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19-fillOutputVectors.R
\name{fillOutputVectors}
\alias{fillOutputVectors}
\title{Fill output vectors on loop completion}
\usage{
fillOutputVectors()
}
\description{
Internal function used to assign
variables to positions corresponding to simulation
run and year in output containers allocated in
\code{\link{defineOutputVectors}}.

Not intended to be called directly, but visible
for the sake of model transparency.
}
