% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish}
\alias{fish}
\title{Growth data for Connecticut River American shad}
\format{A list of 1000 matrices with named dimensions
\describe{
  \item{Sex}{Fish sex}
  \item{Age}{Assigned age}
  \item{Length}{Total length, in cm}
  \item{yearCollected}{Year during which fish was collected}
  \item{backCalculated}{Logical indicating whether or not Length observation was back-calculated}
  \item{Mass}{Mass of fish, in g}
}}
\source{
Connecticut Department of Energy and Environmental Conservation (2015)
}
\usage{
fish
}
\description{
Growth data for spawning American shad
in the Connecticut River, CT, USA.
}
\keyword{datasets}
