% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwFishingMort.R
\name{fwFishingMort}
\alias{fwFishingMort}
\title{Assign freshwater fishing mortality}
\usage{
fwFishingMort()
}
\description{
Internal function used to assign 
freshwater fishing mortality for each PU
in each migration route as an annual rate.

Not intended to be called directly, but visible
for the sake of model transparency.
}
