% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rkm1.R
\name{get_rkm1}
\alias{get_rkm1}
\title{Get rkm1}
\usage{
get_rkm1(river, c_fishAges)
}
\arguments{
\item{river}{One of any rivers implemented in `shadia`.}
}
\description{
Assign first RKM in migration model for each river.
}
