% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaturity.R
\name{getMaturity}
\alias{getMaturity}
\title{Get maturity schedule}
\usage{
getMaturity(region)
}
\arguments{
\item{region}{Life-history region (one of `Semelparous`, 
`Southern Iteroparous` or `Northern Iteroparous`).}
}
\description{
Get maturity schedule by region.
}
