% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaxAge.R
\name{getMaxAge}
\alias{getMaxAge}
\title{Get maximum age}
\usage{
getMaxAge(region)
}
\arguments{
\item{region}{Life-history region (one of `Semelparous`, 
`Southern Iteroparous` or `Northern Iteroparous`).}
}
\description{
Get maximum age for population from ASMFC (2020).
}
