% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hudson_blueback}
\alias{hudson_blueback}
\title{Length and mass of Hudson River American shad}
\format{A dataframe with 2,485 observations of 19 variables
\describe{
  \code{species }{NYSDEC species code}
  
  \code{samp_date }{Date fish was collected}
  
  \code{fl }{Fork length, in mm}
  
  \code{wt }{Mass of fish in grams}
  
  \code{sex }{Sex of fish, `B` is buck (male) and `R`is roe (female)}
  
  \code{age }{Age of fish in years}
  
}}
\source{
New York State Department of Environmental Conservation
}
\usage{
hudson_blueback
}
\description{
Length and mass of spawning American shad
collected in the Hudson River, USA.
}
\keyword{datasets}
