% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hudson_shad}
\alias{hudson_shad}
\title{Length and mass of Hudson River American shad}
\format{A dataframe with 26,564 observations of 4 variables
\describe{
  \code{samp_date }{Date fish was collected}
  
  \code{Length }{Fork length, in mm}
  
  \code{mass }{Mass of fish in grams}
  
  \code{sex }{Sex of fish, `B` is buck (male) and `R`is roe (female)}
}}
\source{
New York State Department of Environmental Conservation
}
\usage{
hudson_shad
}
\description{
Length and mass of spawning American shad
collected in the Hudson River, USA.
}
\keyword{datasets}
