% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualSampling.R, R/inner-loop-sampling.R
\name{innerLoopSampling}
\alias{innerLoopSampling}
\title{Stochastic sampling of annual variables}
\usage{
innerLoopSampling(habitat)

innerLoopSampling(habitat)
}
\description{
Internal function used to perform all
annually based stochastic sampling for the 
individual-based upstream migration model.

Not intended to be called directly, but visible
for the sake of model transparency.

Internal function used to perform all
annually based stochastic sampling for the 
individual-based upstream migration model.

Not intended to be called directly, but visible
for the sake of model transparency.
}
