% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineFunctions.R
\name{invlogit}
\alias{invlogit}
\title{Inverse logit}
\usage{
invlogit(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector on the probability scale [0, 1].
}
\description{
A function for back-transformation 
of variables from the logit scale.
}
