% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-length_weight.R, R/data.R
\docType{data}
\name{length_weight}
\alias{length_weight}
\title{Length-weight regression coefficients for American shad}
\format{
A data frame with 6 observations of 4 variables:
\describe{

\code{region}{ Life-history region}

\code{sex}{ Fish sex (gender)}

\code{alpha}{ Intercept}

\code{beta}{ Slope}
}

A data frame with 6 observations of 4 variables:
\describe{

\code{region}{ Life-history region}

\code{sex}{ Fish sex (gender)}

\code{alpha}{ Intercept}

\code{beta}{ Slope}
}
}
\source{
Atlantic States Marine Fisheries Commission

Atlantic States Marine Fisheries Commission
}
\usage{
length_weight

length_weight
}
\description{
A dataset containing sex-specific regression
coefficients of log10length-log10weight relationships
for each life-history region (NI = Northern iteroparous,
SI = Southern iteroparous, SP = Semelparous). Use to predict
weight at age from length derived using region-specific
vbgf parameters.

A dataset containing sex-specific regression
coefficients of log10length-log10weight relationships
for each life-history region (NI = Northern iteroparous,
SI = Southern iteroparous, SP = Semelparous). Use to predict
weight at age from length derived using region-specific
vbgf parameters.
}
\keyword{datasets}
