% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merrimackRiverModel.R
\name{merrimackRiverModel}
\alias{merrimackRiverModel}
\title{Penobscot River Model}
\usage{
merrimackRiverModel(nRuns = 1, nYears = 50, timing = 1,
  upstream = list(essex = 1, pawtucket = 1, amoskeag = 1, hookset = 1),
  downstream = list(essex = 1, pawtucket = 1, amoskeag = 1, hookset = 1),
  inRiverF = 0, commercialF = 0, bycatchF = 0, indirect = 1,
  latent = 1, watershed = TRUE)
}
\arguments{
\item{nRuns}{The number of times that the
model will be run.}

\item{nYears}{The number of years for which
each run will last. The default is 50 years
to match hydropower license duration in the
Penobscot River.}

\item{timing}{The amount of time required for
upstream passage by individual fish (in days),
where the default (1) indicates a 24-h dam
passage performance standard.}

\item{upstream}{A named list of upstream dam
passage efficiencies at each dam in the
Penobscot River. Stillwater and Orono dams are
not included in the list of values because all
fish reaching Orono Dam are trucked upstream.

Users may specify a single value of upstream
passage at each dam, or a vector of upstream
passage efficiencies at each dam. Note that
passage efficiences passed as vectors are
randomly sampled during each model run
(not each year). Therefore, multiple model runs
are necessary if more than one passage efficiency
is supplied for any dam.}

\item{downstream}{A named list of downstream
dam passage efficiencies at each dam in the
Penobscot River (including Orono and Stillwater
dams).

Users may specify a single value of downstream
passage at each dam, or a vector of downstream
passage efficiencies at each dam. Note that
passage efficiences passed as vectors are
randomly sampled during each model run
(not each year). Therefore, multiple model runs
are necessary if more than one passage efficiency
is supplied for any dam.}

\item{inRiverF}{Annual, recreational harvest of
American shad.
Parameterized as an annual rate [0, 1].}

\item{commercialF}{Commercial fishery mortality
for American shad in marine environment incurred
through targeted fisheries. Parameterized as an
annual rate [0, 1].}

\item{bycatchF}{Marine bycatch mortality of
American shad in non-target fisheries.
Parameterized as an annual rate [0, 1].}

\item{indirect}{Indirect mortality incurred during
freshwater migration as a result of dam-related
impacts (e.g., injury, predation, etc.).}

\item{latent}{Latent mortality incurred during estuary
passage as a result of dam-related impacts (e.g., injury,
delay, etc.).}

\item{watershed}{A logical indicating whether or not
to use the same dam passage efficiencies at all dams
for upstream and downstream. If watershed = TRUE, then
the first element in lists `upstream` and `downstream`
are recycled for all subsequent dams.}
}
\value{
Returns a dataframe of user-defined
inputs and available model outputs for each use that
contains \code{nRuns} x \code{nYears} number of rows.

If run in parallel, returns a list
of dataframes.

The folowing named columns are returned:
\itemize{
    \item \code{year} Year of simulation

    \item \code{time} Passage timing input by user

    \item \code{EssUp...HookUp} User-specified upstream passage efficiencies

    \item \code{EssD...HookD}  User-specified downstream passage efficiencies

    \item \code{pRepeat_Age1...Age11} Age-specific probability of repeat spawning

    \item \code{populationSize} Total number of adult spawners returning to the river

    \item \code{N_I...N_V} Production unit-specific population size after in-river fishery mortality
}
}
\description{
Runs American shad dam passage performance
standard model for Penobscot River, Maine,
USA
}
\section{Warning about serial execution and memory limits}{

Current implementation is based on work
in press, and is thus subject to modification
without notice.

Currently, internal functions rely on \code{list2env} to return
lists to a temporary environment created in the
\code{merrimackRiverModel} function. Consequently, lists
that are exported must be limited in size. Therefore,
users currently need to limit the number of runs per
call (\code{nRuns} argument) to less than 10 or R will
hit memory limits quickly. In reality, serial
execution is prohibitively slow unless implemented
using manual parallel processing (e.g., bash scripting).

In order to achieve a desired number of runs for a given
set of inputs, the recommended approach is to use
parallel execution as demonstrated using snowfall in the
example at the bottom of this page.**COMING SOON**
}

