% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18-nextGeneration.R
\name{nextGeneration}
\alias{nextGeneration}
\title{Project population to the next year}
\usage{
nextGeneration()
}
\description{
Internal function to perform
cohort-based projections of the population
into the next year of the simulation based
on marine survival, commercial fishery harvest,
and fishery bycatch (all annual rates).

Not intended to be called directly, but
visible for the sake of transparency.
}
