% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outer-loop-sampling.R
\name{outerLoopSampling}
\alias{outerLoopSampling}
\title{Stochastic sampling of simulation-specific variables}
\usage{
outerLoopSampling()
}
\description{
Internal function used to perform all
iteration-based stochastic sampling, including
management decisions (passage efficiencies,
timing, etc.).

Not intended to be called directly, but visible
for the sake of model transparency.
}
