% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PenobscotRiverModel.R
\name{penobscotRiverModel}
\alias{penobscotRiverModel}
\title{Penobscot River Model}
\usage{
penobscotRiverModel(nRuns = 1, nYears = 50, timing = 1,
  upstream = list(milford = 1, howland = 1, westEnfield = 1, brownsMill = 1,
  moosehead = 1, guilford = 1, weldon = 1), downstream = list(stillwater = 1,
  orono = 1, milford = 1, howland = 1, westEnfield = 1, brownsMill = 1,
  moosehead = 1, guilford = 1, weldon = 1), pinHarvest = 0, inRiverF = 0,
  commercialF = 0, bycatchF = 0, indirect = 1, latent = 1,
  watershed = TRUE)
}
\arguments{
\item{nRuns}{The number of times that the
model will be run.}

\item{nYears}{The number of years for which
each run will last. The default is 50 years
to match hydropower license duration in the
Penobscot River.}

\item{timing}{The amount of time required for
upstream passage by individual fish (in days), 
where the default (1) indicates a 24-h dam
passage performance standard.}

\item{upstream}{A named list of upstream dam
passage efficiencies at each dam in the 
Penobscot River. Stillwater and Orono dams are
not included in the list of values because all
fish reaching Orono Dam are trucked upstream.

Users may specify a single value of upstream
passage at each dam, or a vector of upstream
passage efficiencies at each dam. Note that
passage efficiences passed as vectors are 
randomly sampled during each model run 
(not each year). Therefore, multiple model runs
are necessary if more than one passage efficiency
is supplied for any dam.}

\item{downstream}{A named list of downstream
dam passage efficiencies at each dam in the 
Penobscot River (including Orono and Stillwater
dams). 

Users may specify a single value of downstream
passage at each dam, or a vector of downstream
passage efficiencies at each dam. Note that
passage efficiences passed as vectors are 
randomly sampled during each model run 
(not each year). Therefore, multiple model runs
are necessary if more than one passage efficiency
is supplied for any dam.}

\item{pinHarvest}{In-river, sustenance harvest
by Penobscot Indian Nation (PIN) upstream of Weldon 
Dam. Parameterized as an annual rate [0, 1].}

\item{inRiverF}{Annual, recreational harvest of 
American shad downstream of Weldon Dam. 
Parameterized as an annual rate [0, 1].}

\item{commercialF}{Commercial fishery mortality
for American shad in marine environment incurred 
through targeted fisheries. Parameterized as an 
annual rate [0, 1].}

\item{bycatchF}{Marine bycatch mortality of
American shad in non-target fisheries. 
Parameterized as an annual rate [0, 1].}

\item{indirect}{Indirect mortality incurred during
freshwater migration as a result of dam-related
impacts (e.g., injury, predation, etc.).}

\item{latent}{Latent mortality incurred during estuary
passage as a result of dam-related impacts (e.g., injury,
delay, etc.).}

\item{watershed}{A logical indicating whether or not
to use the same dam passage efficiencies at all dams
for upstream and downstream. If watershed = TRUE, then
the first element in lists `upstream` and `downstream`
are recycled for all subsequent dams.}
}
\value{
Returns a dataframe of user-defined
inputs and available model outputs for each use that 
contains \code{nRuns} x \code{nYears} number of rows. 

If run in parallel, returns a list
of dataframes.

The folowing named columns are returned:

\code{year} Year of simulation

\code{time} Passage timing input by user

\code{milford_up...guilford_up} User-specified upstream passage efficiencies
 
\code{stillwater_down...weldon_down}  User-specified downstream passage efficiencies
 
\code{pRepeat_Age1...Age9} Age-specific probability of repeat spawning  
 
\code{populationSize} Total number of adult spawners returning to the river

\code{N_pu1A2A...N_pu4B} Production unit-specific population size after in-river fishery mortality
}
\description{
Runs American shad dam passage performance
standard model for Penobscot River, Maine,
USA
}
\examples{

Watershed implementation with passage efficiency
increments of 0.10 and all combinations of upstream
and downstream rates with 24-h standard:

penobscotRiverModel(nRuns = 10,
 upstream = list(milford = seq(0, 1, 0.10),
                 howland = 1,
                 westEnfield = 1,
                 brownsMill = 1,
                 moosehead = 1,
                 guilford = 1,
                 weldon = 1),
 downstream = list(stillwater = seq(0, 1, 0.10),
                   orono = 1,
                   milford = 1,
                   howland = 1,
                   westEnfield = 1,
                   brownsMill = 1,
                   moosehead = 1,
                   guilford = 1,
                   weldon = 1))


Variable upstream and downstream passage at 
each dam in increments of 0.10, with timing of 1 d:

 penobscotRiverModel(nRuns = 10,
  upstream = list(milford = seq(0, 1, 0.10),
                  howland = seq(0, 1, 0.10),
                  westEnfield = seq(0, 1, 0.10),
                  brownsMill = seq(0, 1, 0.10),
                  moosehead = seq(0, 1, 0.10),
                  guilford = seq(0, 1, 0.10),
                  weldon = seq(0, 1, 0.10)
                  ),
  downstream = list(stillwater = seq(0, 1, 0.10),
                    orono = seq(0, 1, 0.10),
                    milford = seq(0, 1, 0.10),
                    howland = seq(0, 1, 0.10),
                    westEnfield = seq(0, 1, 0.10),
                    brownsMill = seq(0, 1, 0.10),
                    moosehead = seq(0, 1, 0.10),
                    guilford = seq(0, 1, 0.10),
                    weldon = seq(0, 1, 0.10)
                    )
 ) 

}
