% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pnr_proj45}
\alias{pnr_proj45}
\title{Projected daily river temperatures (RCP4.5) in 
Penobscot River, USA}
\format{
A dataframe with 54,787 obs. of 5 variables:
\describe{
  \code{Date }{Date for projection}
  
  \code{Eddington_RCP45_temp }{Daily temperature (C) at Eddington}
}
}
\source{
Integrated Systems Ecology Lab, Gulf Of Maine Research Institute
}
\usage{
pnr_proj45
}
\description{
Daily project temperatures in the Penobscot
River, USA for years 1950-2099 under RCP4.5 based on 
historical temperatures from the USGS gauge at Eddington.
}
\keyword{datasets}
