% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-populationMatrices.R
\name{populationMatrices}
\alias{populationMatrices}
\title{Define in-river population matrices}
\usage{
populationMatrices()
}
\description{
Internal function used to create combinations
of production units and fish ages in each of the four
migration routes. Generalized to apply to eggs by females
at each age, too. Later filled in \code{\link{createPUMatrix}}

Not intended to be called directly, but visible
for the sake of model transparency.
}
