% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCohorts.R
\name{processCohorts}
\alias{processCohorts}
\title{Process numbers of fish by age, sex, PU, and migration route}
\usage{
processCohorts()
}
\value{
A list of lists, each holding the numbers of 
males, females, or eggs in each age class in each 
production unit of each of four migration routes 
prior to the application of post spawning mortality
or carrying capacity.
}
\description{
Internal function used to calculate the
numbers of males, females, and eggs spawned in each 
age class in each production unit of each of four
migration routes.

Not intended to be called directly, but visible
for the sake of model transparency.
}
