% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r.parms}
\alias{r.parms}
\title{von Bertalanffy growth parameters for roes}
\format{
A list of 10000 matrices with named dimensions
\describe{
  \code{dim1 }{Sinf, K, t0}

  \code{dim2 }{Estimate, Std Error, t statistic, and p-value}
}
}
\source{
Connecticut Department of Energy and Environmental Conservation (2015)
}
\usage{
r.parms
}
\description{
Bootstrapped regression parameters for
von Bertalanffy growth parameters for models fit to
female American shad in the Connecticut
River, CT, USA.
}
\keyword{datasets}
