% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r.parms_merrimack}
\alias{r.parms_merrimack}
\title{von Bertalanffy growth parameters for roes}
\format{A list of 10000 matrices with named dimensions
\describe{
  \item{\code{dim1}}{Sinf, K, t0}
  
  \item{\code{dim2}}{Estimate, Std Error, t statistic, and p-value}
}}
\source{
U.S. Fish and Wildlife Service (2018)
}
\usage{
r.parms_merrimack
}
\description{
Bootstrapped regression parameters for
von Bertalanffy growth parameters for models fit to
female American shad in the Merrimack River, USA. Uses
data from fish < 3 years of age from the Connecticut
River, USA to calibrate growth curves.
}
\keyword{datasets}
