% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions}
\alias{regions}
\title{Region-system key for von Bertalanffy growth estimates}
\format{A dataframe with 11 obs. of 3 variables:
\describe{
  \code{System }{Stock assessment unit}
  
  \code{Region }{Management region based on life-history strategy}
  
  \code{Thompsonville }{Numeric identifier for region}
}}
\source{
Atlantic States Marine Fisheries Commission
}
\usage{
regions
}
\description{
A key for matching rivers used in the growth 
analysis for the 2020 ASMFC American shad stock assessment. This is
used to query posterior distributions from von Bertalanffy growth
models used to predict fish length at age and incorporate climate
impacts on growth and mortality.
}
\keyword{datasets}
