% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-scalePop.R
\name{scalePop}
\alias{scalePop}
\title{Scale population to speed computations}
\usage{
scalePop(pop, spawningPool, recruitmentPool, scalar)
}
\value{
A list of vectors containing the 1) population size,
2) spawning pool, and 3) recruitment pool, all scaled
according to rules in \code{\link{setScalar}}.
}
\description{
Internal function for rescaling components of the population
based on scalar variable in \code{\link{setScalar}}.

Not intended to be called directly, but visible for
transparency.
}
