% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-setParameters.R
\name{setParameters}
\alias{setParameters}
\title{Set test parameters and environment variables}
\usage{
setParameters()
}
\value{
A list of generic parameters for code benchmarking
and progress monitoring.
}
\description{
Internal function for setting testing parameters
and environment variables.

Not intended to be called directly, but visible
for transparency.
}
