% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-setScalar.R
\name{setScalar}
\alias{setScalar}
\title{Set population scalar for small rivers}
\usage{
setScalar(spawningPool)
}
\description{
Internal function for setting the
value of the population scalar that is used to
dynamically adjust population size to avoid
computational overflow.

Not intended to be called directly, but visible
for transparency.
}
