% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-setupTemperatureData.R
\name{setUpTemperatureData}
\alias{setUpTemperatureData}
\title{Set up data for simulations}
\usage{
setUpTemperatureData(river)
}
\value{
A list of generic parameters for code benchmarking
and progress monitoring.
}
\description{
Internal functions for manipulations to
empirical data used as inputs, and for fitting of
empirical relationships (e.g. discharge regressions).

Not intended to be called directly, but visible
for transparency.
}
\details{
Have retained data manipulation to retain
integrity of raw data files in built-in data sets and
provide transparency in methods.
}
