% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDailyTemperatures.R
\name{simDailyTemperatures}
\alias{simDailyTemperatures}
\title{Simulate (or retrieve) daily temperatures}
\usage{
simDailyTemperatures(mu, climate, river, current_year, n)
}
\arguments{
\item{mu}{a dataframe with columns `val`, `year`, and `day` created 
from setUpTemperatureData() within riverModel() functions. Used to
simulate new daily temperature observations if `climate == current` in
the model call, else temperatures are retrieved from climate projections
in models where these are implemented (Connecticut and Penobscot only)}
}
\description{
Internal function used to simulate daily
temperatures in annualSampling()

Not intended to be called directly, but visible
for the sake of model transparency.
}
