% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGrowth.R
\name{simGrowth}
\alias{simGrowth}
\title{Simulate growth parameters}
\usage{
simGrowth(female = TRUE)
}
\value{
A numeric vector of length 1 containing one sample
of simultaneously estimated VBGF parameters (as a random draw
from fitted model posteriors).
}
\description{
Internal function used to simulate 
sex-specific length at age from sex-specific
von Bertalanffy growth functions (VBGF) incorporating
effect of sst on VBGF parameters L-infinity and K.
}
\details{
This function relies on model estimated VBGF parameters
from the 2020 Atlantic States Marine Fishery Commission 
American shad stock assessment. The VBGF was fit using SST as a 
continuous predictor of parameters Linf and K in the model. We 
use these relationships to predict new values of Linf and K from
projected sea surface temperatures with stored posteriors from
built-in, sex-specific data sets containing VBGF parameters.
}
