% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMarineS.R
\name{simMarineS}
\alias{simMarineS}
\title{Simulate marine survival}
\usage{
simMarineS()
}
\value{
A numeric vector of length 1 containing a single simulated
value of marine survival sampled randomly from a posterior predictive
distribution incorporating projected sea surface temperature 
during the current model year.
}
\description{
Internal function used to simulate marine
survival from posterior distributions of von Bertalanffy
growth function parameters. This function relies on 
prediction of climate-informed vbgf parameter estimates
using sex-aggregated model fit.
}
\details{
This function relies on model estimated VBGF parameters
from the 2020 Atlantic States Marine Fishery Commission 
American shad stock assessment. The VBGF was fit using SST as a 
continuous predictor of parameters Linf and K in the model. We 
use these relationships to predict new values of Linf and K from
projected sea surface temperatures with stored posteriors from
a sex-aggregated data set.
}
