% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStartingPop.R
\name{simStartingPop}
\alias{simStartingPop}
\title{Create starting population}
\usage{
simStartingPop(adults, maxAge, oceanSurvival, spawnRecruit)
}
\value{
A list containing the initial population size,
number of spawners and number of non-spawning individuals
for the first year of simulation.
}
\description{
Internal function for creating a starting
population of fish based on a draw from a negative binomial
calibrated to adult spawning numbers. Number of spawners
is not exact match to `adults`.
}
\examples{
test <- simStartingPop(1e4, 13, 0.45)
sum(test$spawningPool)
plot(test$spawningPool, type='l')

}
