% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTemperature.R
\name{simTemperature}
\alias{simTemperature}
\title{Simulate daily temperatures}
\usage{
simTemperature()
}
\value{
A list containing 1) a matrix of projected 
temperatures for the number of years specified (predTemps), 
and 2) a matrix.
}
\description{
Internal function used to simulate daily
temperatures in each river system using historical gauge
data loaded in \code{setUPData}.

Not currently used in package.
}
\details{
This function relies on a data set (\code{mu}) containing
daily temperatures (\code{val}), year (\code{year}),
and ordinal date (\code{day}). Missing, zero, and negative
values will throw warnings.
}
