% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startingPop.R
\name{startingPop}
\alias{startingPop}
\title{Create starting population}
\usage{
startingPop(river)
}
\value{
A list containing the initial population size,
spawner-recruitment and repeat-spawning probabilities,
and spawning/recruitment pools.
}
\description{
Internal function for creating a starting
population of fish for each iteration.

Not intended to be called directly, but visible
for the sake of model transparency.
}
