% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineFunctions.R
\name{substrRight}
\alias{substrRight}
\title{Substring right}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{A character string or vector of
character strings.}

\item{n}{The number of characters from the right
to collect.}
}
\value{
A character string of n elements from
the right of \code{x}.
}
\description{
Collect characters from the right side of a text string.
}
\examples{

substrRight("shadia", n)
}
