% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tempD}
\alias{tempD}
\title{Connecticut River temperatures, Stratford, NH, USA}
\format{A data frame with 2919 obs. of 7 variables
\describe{
  \item{staid}{USGS Station ID}
  \item{val}{Temperature in degrees celcius}
  \item{dates}{Temperature, in degrees celcius}
  \item{qualcode}{Qualification code}
  \item{year}{date and time of temperature observation}
  \item{day}{Temperature, in degrees celcius}
  \item{ph}{Photoperiod, in hours}
}}
\source{
U.S. Geological Survey
}
\usage{
tempD
}
\description{
Temperature data from Stratford, NH, USA.
}
\keyword{datasets}
