% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tempData}
\alias{tempData}
\title{Penobscot River temperatures, 
Eddington, ME, USA}
\format{A data frame with 13265 obs. of 6 variables
\describe{
  \item{staid}{USGS Station ID}
  \item{val}{Temperature in degrees celcius}
  \item{dates}{Temperature, in degrees celcius}
  \item{qualcode}{Qualification code}
  \item{year}{date and time of temperature observation}
  \item{day}{Temperature, in degrees celcius}
}}
\source{
U.S. Geological Survey
}
\usage{
tempData
}
\description{
Temperature data for 
Penobscot River temperatures, Eddington, ME, USA
}
\keyword{datasets}
