% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tempData_androscoggin}
\alias{tempData_androscoggin}
\title{Androscoggin River Temperatures}
\format{
A data frame with 3,008 observations of 5 variables
\describe{
  \code{staid }{Site ID in watershed}

  \code{dates }{Date of temperature observation}

  \code{year }{Year of temperature observation}

  \code{day }{Ordinal date}

  \code{val }{Mean daily bottom temperature in degrees celcius}
}
}
\source{
SHEDS Stream Temperature Database
}
\usage{
tempData_androscoggin
}
\description{
Temperature data from
various monitoring sites in the
Androscoggin and Little Androscoggin 
rivers, Maine, USA.
}
\keyword{datasets}
