% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tempData_connecticut}
\alias{tempData_connecticut}
\title{Connecticut River temperatures,
Turner's Falls, MA, USA.}
\format{
A data frame with 167,333 obs. of 3 variables
\describe{
  \code{val }{Temperature in degrees celcius}
  
  \code{year }{Year of temperature observation}
  
  \code{day }{Ordinal date}
}
}
\source{
T. Castro-Santos, U.S. Geological Survey
}
\usage{
tempData_connecticut
}
\description{
Sub-daily temperatures from USGS S.O.
Conte Anadromous Fish Lab in Turner's Falls, MA
for the 21-year period 1994-2015.
}
\keyword{datasets}
