% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tempData_hudson}
\alias{tempData_hudson}
\title{Hudson River Temperatures}
\format{A data frame with 4782 observations of 6 variables
\describe{
  \code{staid }{Site ID in watershed}
  
  \code{val }{Temperature in degrees celcius}
  
  \code{dates }{Date of temperature observation}
  
  \code{qualcode }{Quality control code}
  
  \code{day }{Ordinal date}
  
  \code{year }{Year of temperature observation}
}}
\source{
U.S. Geological Survey
}
\usage{
tempData_hudson
}
\description{
Temperature data from the Hudson River near
Albany, NY, USA (USGS Gage ID 01359139).
}
\keyword{datasets}
