% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tempData_kennebec}
\alias{tempData_kennebec}
\title{Kennebec River Temperatures}
\format{A data frame with 20808 observations of 5 variables
\describe{
  \code{staid }{Site ID in watershed}
  
  \code{dates }{Date of temperature observation}
  
  \code{year }{Year of temperature observation}
  
  \code{day }{Ordinal date}
  
  \code{val }{Bottom temperature in degrees celcius}
}}
\source{
Maine Department of Marine Resources, SHEDS Stream Temperature Database
}
\usage{
tempData_kennebec
}
\description{
Temperature data from 
various monitoring sites in the 
Kennebec River, Maine, USA.
}
\keyword{datasets}
