% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tempData_susquehanna}
\alias{tempData_susquehanna}
\title{Susquehanna River at Danville, PA}
\format{
A data frame with 1816 observations of 6 variables
\describe{
  \code{staid }{USGS Station ID}

  \code{dates }{Date of temperature observation}

  \code{val }{Temperature in degrees celcius}

  \code{day }{Ordinal date}
  
  \code{year }{Year of temperature observation}
}
}
\source{
U.S. Geological Survey
}
\usage{
tempData_susquehanna
}
\description{
Temperature data for
the Susquehanna River at Danville, PA.
}
\keyword{datasets}
