% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-01-upstreamMortHudson.R
\name{upstreamMort}
\alias{upstreamMort}
\title{Assign upstream mortality during upstream
passage of dams.}
\usage{
upstreamMort(upstreamMortality, river, nRoutes)
}
\description{
Internal function used to assign
cumulative upstream mortality incurred during passage
of locks.

Only implemented in \code{mohawkHudsonRiverModel()}
A single value is applied to all dams in the 
watershed for now.

Not intended to be called directly, but visible
for the sake of model transparency.
}
