% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vbgf_agg}
\alias{vbgf_agg}
\title{Sex-aggregate von Bertalanffy growth parameters}
\format{
A list with 7 elements, each of which corresponds to an
estimated parameter in the sex-aggregate model described in Gilligan (In prep).
}
\source{
Gilligan, E. (In Prep). MS Biology Thesis, SUNY Oneonta.
}
\usage{
vbgf_agg
}
\description{
Posterior distributions for parameters of a 
sex-aggregated von Bertalanffy growth function for American
shad in the Northeast Continental Shelf Large Marine Ecosystem.
}
\keyword{datasets}
