% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vbgf_f}
\alias{vbgf_f}
\title{Female von Bertalanffy growth parameters for
American shad}
\format{
A list with 7 elements, each of which corresponds to an
estimated parameter in the female-only model described in Gilligan (In prep).
}
\source{
Gilligan, E. (In Prep). MS Biology Thesis, SUNY Oneonta.
}
\usage{
vbgf_f
}
\description{
Posterior distributions for parameters of a
female-only von Bertalanffy growth function for American
shad in the Northeast Continental Shelf Large Marine Ecosystem.
}
\keyword{datasets}
