% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vbgf_hudson_f}
\alias{vbgf_hudson_f}
\title{Female von Bertalanffy growth parameters for
blueback herring in the Hudson River}
\format{
A list with 3 elements, each of which corresponds to an
estimated parameter in the model (Linf, K, t0).
}
\source{
New York State Department of Environmental Conservation, unpublished data
}
\usage{
vbgf_hudson_f
}
\description{
Posterior distributions for parameters of a
female-only von Bertalanffy growth function for blueback herring
in the Hudson River, NY, USA.
}
\keyword{datasets}
