% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vbgf_kennebec_bbh_agg}
\alias{vbgf_kennebec_bbh_agg}
\title{Sex-aggregate von Bertalanffy growth parameters for
blueback herring in the Kennebec River}
\format{
A list with elements corresponding to estimated parameters
of a von Bertalanffy growth function with sex aggregated
}
\source{
Maine Department of Marine Resources, unpublished
}
\usage{
vbgf_kennebec_bbh_agg
}
\description{
Posterior distributions for parameters of a
sex-aggregated von Bertalanffy growth function for blueback herring
in the Kennebec River, Maine, USA.
}
\keyword{datasets}
