% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vbgf_kennebec_bbh_m}
\alias{vbgf_kennebec_bbh_m}
\title{Female von Bertalanffy growth parameters for
blueback herring in the Hudson River}
\format{
A list with elements corresponding to estimated parameters
of a von Bertalanffy growth function for males
}
\source{
Maine Department of Marine Resources, unpublished
}
\usage{
vbgf_kennebec_bbh_m
}
\description{
Posterior distributions for parameters of a 
male-only von Bertalanffy growth function for blueback herring
in the Kennebec River, Maine, USA.
}
\keyword{datasets}
