% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weldonScenarios.R
\name{weldonScenarios}
\alias{weldonScenarios}
\title{Weldon Dam relicensing scenarios in Penobscot River}
\usage{
weldonScenarios()
}
\value{
A list containing the initial population size,
spawner-recruitment and repeat-spawning probabilities,
and spawning/recruitment pools.
}
\description{
Special, internal function to replace Weldon passage
depending on when passage is implemented. This is a relic from
specific management strategy evaluation scenarios related to
hydropower relicensing, which is not implemented in the
current release. Retained as a reminder of flexibility, and for
potential generalized usage in future releases.

Not intended to be called directly, but visible
for the sake of model transparency.
}
