% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeData.R
\name{writeData}
\alias{writeData}
\title{Write simulation results}
\usage{
writeData()
}
\value{
A list of results.
}
\description{
Internal function used to collect and
return all model inputs and relevant model
outputs (e.g., population size) from all k iterations 
of n years to the temporary environment 
in \code{\link{penobscotRiverModel}}, 
\code{\link{merrimackRiverModel}}, and
\code{\link{connecticutRiverModel}}.

Not intended to be called directly, but visible 
nonetheless.
}
