% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish}
\alias{fish}
\title{Growth data for Connecticut River American shad}
\format{
A dataframe with 19,946 observatoins of 6 variables
\describe{
  \code{Sex }{Fish sex}
  
  \code{Age }{Assigned age}
  
  \code{Length }{Total length, in cm}
  
  \code{yearCollected }{Year during which fish was collected}
  
  \code{backCalculated }{Logical indicating whether or not Length observation was back-calculated}
  
  \code{Mass }{Mass of fish, in g}
}
}
\source{
Connecticut Department of Energy and Environmental Conservation (2015)
}
\usage{
fish
}
\description{
Growth data for spawning American shad
in the Connecticut River, CT, USA.
}
\keyword{datasets}
