% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-03-getPhotoperiod.R
\name{getPhotoperiod}
\alias{getPhotoperiod}
\title{Get maturity schedule}
\usage{
getPhotoperiod(river, day)
}
\arguments{
\item{river}{One of any rivers implemented in `shadia`.}
}
\description{
Calculate photoperiod by river.
}
