% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineFunctions.R
\name{processPopulation}
\alias{processPopulation}
\title{Process population}
\usage{
processPopulation(isFemale, isEgg = FALSE)
}
\description{
A function that uses \code{\link{createPUMatrix}},
\code{\link{assignFishToRoutes}}, and 
\code{\link{additionalEggsProcessing}} to return the 
number of fish in each cohort in each PU prior to 
application of annual mortality rates and downstream
migration. Not intended to be called directly.
 Visible for the sake of model transparency.
}
