% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rcp45_sst}
\alias{rcp45_sst}
\title{Projected sea surface temperatures RCP 4.5}
\format{A dataframe with 118 obs. of 5 variables:
\describe{
  \code{Year }{Year for projection}
  
  \code{Scenario }{Representative climate pathway (RCP) used for projection}
  
  \code{Mean }{Mean for projection}

  \code{pct5th }{5th percentile for projection}

  \code{pct95th }{95th percentile for projection}                              
}}
\source{
Integrated Systems Ecology Lab, Gulf Of Maine Research Institute
}
\usage{
rcp45_sst
}
\description{
Projected sea surface temperatures for
the Northeast Continental Shelf Large Marine Ecosystem
(NELME) under RCP 4.5 scenario for years 1981 - 2099.
}
\keyword{datasets}
