% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tempD}
\alias{tempD}
\title{Connecticut River temperatures, Stratford, NH, USA}
\format{A data frame with 2919 obs. of 7 variables
\describe{
  \item{\code{staid}}{USGS Station ID}
  
  \item{\code{val}}{Temperature in degrees celcius}
  
  \item{\code{dates}}{Date of measurement}
  
  \item{\code{qualcode}}{Qualification code}
  
  \item{\code{year}}{Year of temperature observation}
  
  \item{\code{day}}{Ordinal date}
  
  \item{\code{ph}}{Photoperiod, in hours}
}}
\source{
U.S. Geological Survey
}
\usage{
tempD
}
\description{
Temperature data from Stratford, NH, USA.
}
\keyword{datasets}
