% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tempD}
\alias{tempD}
\title{Connecticut River temperatures, Stratford, NH, USA}
\format{A data frame with 2919 obs. of 7 variables
\describe{
  \code{staid }{USGS Station ID}
  
  \code{val }{Temperature in degrees celcius}
  
  \code{dates }{Date of measurement}
  
  \code{qualcode }{Qualification code}
  
  \code{year }{Year of temperature observation}
  
  \code{day }{Ordinal date}
  
  \code{ph }{Photoperiod, in hours}
}}
\source{
U.S. Geological Survey
}
\usage{
tempD
}
\description{
Temperature data from Stratford, NH, USA.
}
\keyword{datasets}
