% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tempData}
\alias{tempData}
\title{Penobscot River temperatures, 
Eddington, ME, USA}
\format{A data frame with 13265 obs. of 6 variables
\describe{
  \item{\code{staid}}{USGS Station ID}
  
  \item{\code{val}}{Temperature in degrees celcius}
  
  \item{\code{dates}}{Date of temperature observation}
  
  \item{\code{qualcode}}{Qualification code}
  
  \item{\code{year}}{Year of temperature observation}
  
  \item{\code{day}}{Ordinal date}
}}
\source{
U.S. Geological Survey
}
\usage{
tempData
}
\description{
Temperature data for 
Penobscot River temperatures, Eddington, ME, USA
}
\keyword{datasets}
