% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tempData_hudson}
\alias{tempData_hudson}
\title{Kennebec River Temperatures}
\format{A data frame with 5884 observations of 5 variables
\describe{
  \code{staid }{Site ID in watershed}
  
  \code{val }{Temperature in degrees celcius}
  
  \code{dates }{Date of temperature observation}
  
  \code{qualcode }{Quality control code}
  
  \code{day }{Ordinal date}
  
  \code{year }{Year of temperature observation}
}}
\source{
U.S. Geological Survey
}
\usage{
tempData_hudson
}
\description{
Temperature data from 
various monitoring sites in the 
Kennebec River, Maine, USA.
}
\keyword{datasets}
