% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tempData_hudson}
\alias{tempData_hudson}
\title{Hudson River Temperatures}
\format{A data frame with 4787 observations of 5 variables
\describe{
  \code{staid }{Site ID in watershed}
  
  \code{val }{Temperature in degrees celcius}
  
  \code{year }{Year of temperature observation}

  \code{dates }{Date of temperature observation}
  
  \code{day }{Ordinal date}
  
}}
\source{
Maine Department of Marine Resources
}
\usage{
tempData_hudson
}
\description{
Temperature data from 
USGS gage 01359139 monitoring sites in the 
Hudson River, NY, USA.
}
\keyword{datasets}
