% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tempData_susquehanna}
\alias{tempData_susquehanna}
\title{Saco River at River Street in Bartlett, NH}
\format{
A data frame with 1816 observations of 6 variables
\describe{
  \code{staid }{USGS Station ID}
  
  \code{val }{Temperature in degrees celcius}
  
  \code{dates }{Date of temperature observation}
  
  \code{qualcode }{Qualification code}
  
  \code{year }{Year of temperature observation}
  
  \code{day }{Ordinal date}
}
}
\source{
U.S. Geological Survey
}
\usage{
tempData_susquehanna
}
\description{
Temperature data for 
the Susquehanna River at Harrisburg, PA.
}
\keyword{datasets}
